+++ cmd/ask.go
@@ -41,6 +41,11 @@ var askCmd = &cobra.Command{
 			os.Exit(1)
 		}
 
+		flags := []flagFunc{scoreThresholdFlag, maxResultsFlag}
+		for _, v := range flags {
+			v(cmd)
+		}
+
 		bq := bhlquestFactory()
 		q := args[0]
 		answer, err := bq.Ask(q)
@@ -57,13 +62,12 @@ var askCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(askCmd)
 
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// askCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// askCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	askCmd.Flags().IntP(
+		"max-results", "m", 0,
+		"defines the maximum number of results",
+	)
+	askCmd.Flags().Float64P(
+		"score-threshold", "s", 0,
+		"results with lower score will be ignored",
+	)
 }