+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add parameters for maximum number of returns and score threshold (#13)
+- Change limiting factor from main_taxa to main_class of an item (#12)
 - Add better hanling of initiation process (#11)
 - Add text of matched chunk to the output (if possible) (#10)
 - Add an option to limit BHL items intake by a list of taxa (#9)
@@ -30,9 +32,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This document follows [changelog guidelines]
 
+[v0.0.2]: https://github.com/gnames/bhlquest/compare/v0.0.1...v0.0.2
 [v0.0.1]: https://github.com/gnames/bhlquest/compare/v0.0.0...v0.0.1
 [v0.0.0]: https://github.com/gnames/bhlquest/tree/v0.0.0
 
+[#20]: https://codeberg.org/dimus/madcow/issues/20
+[#19]: https://codeberg.org/dimus/madcow/issues/19
+[#18]: https://codeberg.org/dimus/madcow/issues/18
+[#17]: https://codeberg.org/dimus/madcow/issues/17
+[#16]: https://codeberg.org/dimus/madcow/issues/16
+[#15]: https://codeberg.org/dimus/madcow/issues/15
+[#14]: https://codeberg.org/dimus/madcow/issues/14
+[#13]: https://codeberg.org/dimus/madcow/issues/13
 [#12]: https://codeberg.org/dimus/madcow/issues/12
 [#11]: https://codeberg.org/dimus/madcow/issues/11
 [#10]: https://codeberg.org/dimus/madcow/issues/10